// landing/fragments/landing-price/src/price.interfaces.ts
export interface PriceProps {
  costPerMonth: string
  currency: string
  fullCost: string
  economy: string
  quantityVideoLessons: string
  firstLineCircle: string
  secondLineCircle: string
  quantityMonths: string
  backgroundRectangle: string
  squareRotate: number
  circleRotate: number
  rectangleRotate: number
  rectanglePositionX: number
  rectanglePositionY: number
}